requirements.txt
langchain
langchain-groq
langgraph
pandas
plotly
streamlit
python-dotenv
tabulate
openpyxl
